Derivation notes

For a series of Bernoulli obervations, Y_i, the variance of the observations is given by

Usage

.sd.y(model)

Details

$$Var(Y) = (( Y_i - _Y )^2)n-1$$

where $_Y$ is the mean of the observations.

However, if the data is structured as Binomial observations, $P_j = k_jn_j$ ($n_j$ trials, $k_j$ successes, and $r_j$ failures) we need to weight the observations by the number of trials, $n_j$ to understand the underlying Bernoulli process.

First, to determine the mean of the underlying Bernoulli process:

$$_y = _i Y_in = _j k_jn$$

i.e. the sum of observed successes divided by the total number of trials.

Then, we can sub-divide the sum of squares into the sum of squares of the successes and the sum of squares of the failures:

$$Var(Y) = (( Y_i - _Y )^2)n-1 $$ $$ = _j k_j (1-_Y)^2 + r_j (0-_Y)^2n-1 $$ $$ = _j k_j (1-_Y)^2+ r_j _Y^2n-1 $$

We follow simialr process to enumerate the variance of the fitted values, $y$. The mean of the fitted values is:

$$_y = _j n_j y_jn$$

and the variance is:

$$Var(Y) = _j n_j (yj - y)^2n-1$$